backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["crazyboy9103 <gil9103@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
bcrypt = "^4.0.1"
click = "^8.1.3"
fastapi = {extras = ["asyncmy"], version = "^0.95.0"}
pydantic = {extras = ["email"], version = "^1.10.7"}
tortoise-orm = "^0.19.3"
uvicorn = "^0.21.1"
python-dotenv = "^1.0.0"
python-multipart = "^0.0.6"
PyYAML = "^6.0"
aiomysql = "^0.1.1"
passlib = "^1.7.4"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
